Hibernate ORM DevUI - Move dev-console-specific code to a dedicated package
