+++ extensions/hibernate-orm/deployment/src/main/java/io/quarkus/hibernate/orm/deployment/devconsole/DevConsoleProcessor.java
@@ -3,7 +3,7 @@ package io.quarkus.hibernate.orm.deployment.devconsole;
 import io.quarkus.deployment.IsDevelopment;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.devconsole.spi.DevConsoleRuntimeTemplateInfoBuildItem;
-import io.quarkus.hibernate.orm.runtime.PersistenceUnitInfoSupplier;
+import io.quarkus.hibernate.orm.runtime.devconsole.PersistenceUnitInfoSupplier;
 
 public class DevConsoleProcessor {
